Add files via upload
Patients with Liver disease have been continuously increasing because of excessive consumption of alcohol, inhale of harmful gases, intake of contaminated food, pickles and drugs. This data set was used to evaluate prediction algorithms in an effort to reduce burden on doctors.Use these patient records to determine which patients have liver disease and which ones do not.